Workaround for SQLite.JS handling of BigInt in query parameters.
SQLite.JS binds bigint as a string. This was fine for insert/update insertions, however causes unexpected behavior on comparisons.

Eventually the implementation should probably shift to the new official SQLite WASM. In the meantime, as a workaround, query parameters that are known to be long on Kotlin/JS are automatically converted to CAST(? AS BIGINT).