Avoid potential invalidation tracker deadlock:
When an onInvalidated callback ran a blocking query, this could cause a deadlock when using SQLite which uses one conenction at a time (the callback won't finish until a connection is available, but the connection won't be available until the useNewConnectionInternal function calling the callback finishes).

Make invalidationTracker callback async using the JdbcImplHelper scope.